refactor: extract SQS sub-components into dedicated modules
Split SQSBrowser.tsx (3100 lines) into 12 focused sub-components
under sqs/ directory: sheets, confirm dialogs, queue card, badges,
pagination, and shared utilities. Deduplicate queue card JSX into
a single QueueCard component with showExtendedStats prop.
