+++ ui/src/components/service-views/sqs/utils.ts
@@ -1,0 +1,15 @@
+export function formatNumber(num: number): string {
+  if (num === 0) return '0'
+  if (num < 1000) return String(num)
+  if (num < 1000000) return `${(num / 1000).toFixed(1)}K`
+  return `${(num / 1000000).toFixed(1)}M`
+}
+
+export function formatDuration(seconds: number): string {
+  if (seconds < 60) return `${seconds}s`
+  if (seconds < 3600) return `${Math.floor(seconds / 60)}m`
+  if (seconds < 86400) return `${Math.floor(seconds / 3600)}h`
+  return `${Math.floor(seconds / 86400)}d`
+}
+
+export const PAGE_SIZE_OPTIONS = [10, 25, 50, 100] as const
